Consolidate newline tables into a shared module
